Move `Location` out of pipeline
Move `Rank`, `Location` and ranks out of the pipeline directory and into
a standalone `Ranks` directory.
