+++ include/revng/Clift/Helpers.h
@@ -7,8 +7,8 @@
 #include <optional>
 
 #include "revng/Clift/Clift.h"
-#include "revng/Pipeline/Location.h"
-#include "revng/Pipes/Ranks.h"
+#include "revng/Ranks/Location.h"
+#include "revng/Ranks/Ranks.h"
 #include "revng/Support/MetaAddress.h"
 
 namespace clift {
