+++ include/revng/CliftImportModel/CAttributeListBuilder.h
@@ -13,8 +13,8 @@
 #include "revng/Clift/Clift.h"
 #include "revng/Clift/CliftAttributes.h"
 #include "revng/PTML/CAttributes.h"
-#include "revng/Pipeline/Location.h"
-#include "revng/Pipes/Ranks.h"
+#include "revng/Ranks/Location.h"
+#include "revng/Ranks/Ranks.h"
 
 namespace clift {
