+++ include/revng/PTML/CBuilder.h
@@ -16,8 +16,8 @@
 #include "revng/PTML/Doxygen.h"
 #include "revng/PTML/IndentedOstream.h"
 #include "revng/PTML/Tag.h"
-#include "revng/Pipeline/Location.h"
-#include "revng/Pipes/Ranks.h"
+#include "revng/Ranks/Location.h"
+#include "revng/Ranks/Ranks.h"
 
 namespace ptml {
