+++ include/revng/Pipeline/Kind.h
@@ -7,8 +7,8 @@
 #include <set>
 
 #include "revng/Pipeline/GlobalTupleTreeDiff.h"
-#include "revng/Pipeline/Rank.h"
 #include "revng/Pipeline/RegisterKind.h"
+#include "revng/Ranks/Rank.h"
 #include "revng/Support/Assert.h"
 
 namespace pipeline {
