+++ include/revng/Pipes/IRHelpers.h
@@ -8,12 +8,9 @@
 
 #include "llvm/ADT/DenseMap.h"
 
-#include "revng/Pipeline/Location.h"
-#include "revng/Pipes/Ranks.h"
+#include "revng/Ranks/Location.h"
+#include "revng/Ranks/Ranks.h"
 #include "revng/Support/MetaAddress.h"
-
-std::optional<pipeline::Location<decltype(revng::ranks::Instruction)>>
-getLocation(const llvm::Instruction *I);
 
 [[nodiscard]] llvm::DenseMap<MetaAddress, const llvm::Function *>
 getTargetToFunctionMapping(const llvm::Module &M);
@@ -22,14 +19,3 @@ namespace llvm {
 class DebugLoc;
 class Instruction;
 } // namespace llvm
-
-// This ensures debug information validity.
-//
-// In revng modules, valid debug information location is one that is:
-// - non-empty (`bool(DebugLoc)`),
-// - has a scope (`DebugLoc->getScope()`) with a non-empty name,
-// - where the name is a valid `/instruction/...` location.
-// (the last one is subject to change when we start attaching more than one
-//  address per llvm instruction).
-llvm::Error isDebugLocationInvalid(const llvm::DebugLoc &Instruction);
-llvm::Error isDebugLocationInvalid(const llvm::Instruction &Instruction);
