+++ include/revng/Ranks/Location.h
@@ -16,7 +16,7 @@
 
 #include "revng/ADT/CompilationTime.h"
 #include "revng/ADT/STLExtras.h"
-#include "revng/Pipeline/Rank.h"
+#include "revng/Ranks/Rank.h"
 #include "revng/Support/MetaAddress.h"
 #include "revng/Support/YAMLTraits.h"
