+++ lib/CliftImportModel/Verify.cpp
@@ -11,8 +11,8 @@
 #include "revng/Model/RawFunctionDefinition.h"
 #include "revng/Model/TypeDefinition.h"
 #include "revng/PTML/CAttributes.h"
-#include "revng/Pipeline/Location.h"
-#include "revng/Pipes/Ranks.h"
+#include "revng/Ranks/Location.h"
+#include "revng/Ranks/Ranks.h"
 
 namespace ranks = revng::ranks;
