+++ lib/CliftPipes/ImportTypes.cpp
@@ -7,10 +7,10 @@
 #include "revng/CliftImportModel/ImportModel.h"
 #include "revng/CliftPipes/CliftContainer.h"
 #include "revng/CliftPipes/ImportTypes.h"
-#include "revng/Pipeline/Location.h"
 #include "revng/Pipeline/RegisterPipe.h"
 #include "revng/Pipes/FileContainer.h"
 #include "revng/Pipes/Kinds.h"
+#include "revng/Ranks/Location.h"
 
 //
 // Old style pipes
