+++ lib/EditCBody/EditCBodyAnalysis.cpp
@@ -22,8 +22,8 @@
 #include "revng/Model/GotoLabel.h"
 #include "revng/Model/LocalVariable.h"
 #include "revng/Model/StatementComment.h"
-#include "revng/Pipeline/Location.h"
-#include "revng/Pipes/Ranks.h"
+#include "revng/Ranks/Location.h"
+#include "revng/Ranks/Ranks.h"
 #include "revng/Support/Debug.h"
 #include "revng/Support/MetaAddress.h"
 #include "revng/Support/MetaAddress/YAMLTraits.h"
