+++ lib/EditCType/EditCTypeAnalysis.cpp
@@ -37,7 +37,8 @@
 #include "revng/Pipeline/Option.h"
 #include "revng/Pipeline/RegisterAnalysis.h"
 #include "revng/Pipes/ModelGlobal.h"
-#include "revng/Pipes/Ranks.h"
+#include "revng/Ranks/Ranks.h"
+#include "revng/Support/PathList.h"
 #include "revng/Support/TemporaryFile.h"
 #include "revng/Support/YAMLTraits.h"
 #include "revng/TupleTree/TupleTreeDiff.h"
