+++ lib/EditCType/EditCTypeHelpers.h
@@ -12,10 +12,10 @@
 #include "revng/Model/TypeDefinition.h"
 #include "revng/Pipeline/Context.h"
 #include "revng/Pipeline/Kind.h"
-#include "revng/Pipeline/Location.h"
 #include "revng/Pipeline/Option.h"
 #include "revng/Pipeline/RegisterAnalysis.h"
 #include "revng/Pipes/ModelGlobal.h"
+#include "revng/Ranks/Location.h"
 #include "revng/Support/Assert.h"
 #include "revng/Support/Debug.h"
 #include "revng/Support/YAMLTraits.h"
