+++ lib/PromoteStackPointer/Helpers.h
@@ -4,7 +4,8 @@
 // This file is distributed under the MIT License. See LICENSE.md for details.
 //
 
-#include "revng/Pipes/IRHelpers.h"
+#include "revng/Model/FunctionTags.h"
+#include "revng/Ranks/IRHelpers.h"
 #include "revng/Support/MetaAddress.h"
 
 inline MetaAddress getCallerBlockAddress(llvm::Instruction *I) {
