+++ lib/Yield/Assembly/LLVMTagsToPTML.cpp
@@ -3,8 +3,8 @@
 //
 
 #include "revng/PTML/Tag.h"
-#include "revng/Pipeline/Location.h"
-#include "revng/Pipes/Ranks.h"
+#include "revng/Ranks/Location.h"
+#include "revng/Ranks/Ranks.h"
 #include "revng/Yield/Function.h"
 
 //
