+++ lib/Yield/HexDump.cpp
@@ -14,10 +14,10 @@
 #include "revng/Model/RawBinaryView.h"
 #include "revng/PTML/Tag.h"
 #include "revng/Pipeline/AllRegistries.h"
-#include "revng/Pipeline/Location.h"
 #include "revng/Pipes/FileContainer.h"
 #include "revng/Pipes/Kinds.h"
-#include "revng/Pipes/Ranks.h"
+#include "revng/Ranks/Location.h"
+#include "revng/Ranks/Ranks.h"
 #include "revng/Support/IRHelpers.h"
 #include "revng/Support/MetaAddress/IntervalContainers.h"
 #include "revng/Yield/HexDump.h"
