+++ lib/Yield/PTML.cpp
@@ -13,8 +13,8 @@
 #include "revng/PTML/CommentPlacementHelper.h"
 #include "revng/PTML/Constants.h"
 #include "revng/PTML/Tag.h"
-#include "revng/Pipeline/Location.h"
-#include "revng/Pipes/Ranks.h"
+#include "revng/Ranks/Location.h"
+#include "revng/Ranks/Ranks.h"
 #include "revng/Yield/ControlFlow/FallthroughDetection.h"
 #include "revng/Yield/Function.h"
 #include "revng/Yield/PTML.h"
