+++ lib/Yield/Pipes/CallGraphPipes.cpp
@@ -5,7 +5,6 @@
 #include "revng/EarlyFunctionAnalysis/ControlFlowGraph.h"
 #include "revng/EarlyFunctionAnalysis/ControlFlowGraphCache.h"
 #include "revng/Model/Binary.h"
-#include "revng/Pipeline/Location.h"
 #include "revng/Pipeline/Pipe.h"
 #include "revng/Pipeline/RegisterContainerFactory.h"
 #include "revng/Pipeline/RegisterPipe.h"
@@ -13,6 +12,7 @@
 #include "revng/Pipes/ModelGlobal.h"
 #include "revng/Pipes/StringMap.h"
 #include "revng/Pipes/TupleTreeContainer.h"
+#include "revng/Ranks/Location.h"
 #include "revng/TupleTree/TupleTree.h"
 #include "revng/Yield/CrossRelations/CrossRelations.h"
 #include "revng/Yield/Generated/ForwardDecls.h"
